requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pydantic==2.8.2
python-jose==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
boto3==1.35.20
python-dotenv==1.0.1


